+++ scvi/module/_vae.py
@@ -326,9 +326,6 @@ class VAE(BaseModuleClass):
 
     @auto_move_data
     def inference_no_encode(self, qz_m, qz_v, latent_data_type, n_samples=1):
-        """
-        Runs the inference routine but skips encoding, used in latent mode
-        """
         assert latent_data_type in ["sampled", "dist"]
         if latent_data_type == "sampled":
             if n_samples > 1:
