+++ scvi/module/base/_base_module.py
@@ -205,9 +205,7 @@ class BaseModuleClass(nn.Module):
 
     @abstractmethod
     def inference_no_encode(self, qz_m, qz_v, latent_data_type, n_samples=1):
-        """
-        Runs the inference routine but skips encoding, used in latent mode
-        """
+        """Runs the inference routine but skips encoding, used in latent mode"""
         pass
 
     @abstractmethod
